Sharsha/appeals 43987.amybids (#22957)
* some working tests

* APPEALS-43987. fix the last jest failure

* APPEALS-43987. cleaup

---------
